refactor: 제안받은 돈길 바텀시트 관련 props 네번 내리는 로직 전역관리 바텀시트로 개선
