+++ src/components/home/proposed/ProposedDongilItem.tsx
@@ -1,23 +1,26 @@
 import InterestBadge from '@components/common/badges/InterestBadge';
 import { modals } from '@components/common/modals/Modals';
+import useAxiosPrivate from '@lib/hooks/auth/useAxiosPrivate';
+import useGlobalBottomSheet from '@lib/hooks/useGlobalBottomSheet';
 import useModals from '@lib/hooks/useModals';
 import { IDongil } from '@lib/types/IDongil';
-import { Dispatch, SetStateAction } from 'react';
+import { TFetchStatus } from '@lib/types/TFetchStatus';
+import { useAppDispatch, useAppSelector } from '@store/app/hooks';
+import { selectSelectedKid } from '@store/slices/kidsSlice';
+import {
+  approveProposedDongil,
+  selectProposedDongils,
+} from '@store/slices/proposedDongilsSlice';
+import { appendThisWeekSDongil } from '@store/slices/thisWeekSDongilsSlice';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 
 interface ProposedDongilItemProps {
   proposedDongil: IDongil;
-  onApproveCheckOpen: () => void;
-  setIdToApprove: Dispatch<SetStateAction<number>>;
 }
 
-function ProposedDongilItem({
-  proposedDongil,
-  onApproveCheckOpen,
-  setIdToApprove,
-}: ProposedDongilItemProps) {
-  const { openModal } = useModals();
+function ProposedDongilItem({ proposedDongil }: ProposedDongilItemProps) {
   const {
     id,
     createdAt,
@@ -28,8 +31,86 @@ function ProposedDongilItem({
     totalPrice,
     weekPrice,
     weeks,
+    fileName,
   } = proposedDongil;
 
+  const {
+    isOpen,
+    setOpenBottomSheet,
+    setCloseBottomSheet,
+    openSheetBySequence,
+  } = useGlobalBottomSheet();
+
+  const selectedKid = useAppSelector(selectSelectedKid);
+  const proposedDongils = useAppSelector(selectProposedDongils);
+  const dispatch = useAppDispatch();
+  const axiosPrivate = useAxiosPrivate();
+  const [approveProposedDongilStatus, setApproveProposedDongilStatus] =
+    useState<TFetchStatus>('idle');
+  const canApproveProposedDongil =
+    approveProposedDongilStatus === 'idle' && selectedKid !== null;
+
+  // 3. 돈길 수락
+  async function handleApproveButtonClick() {
+    if (canApproveProposedDongil) {
+      try {
+        setApproveProposedDongilStatus('pending');
+        await dispatch(
+          approveProposedDongil({
+            axiosPrivate,
+            idToApprove: id,
+            isApprove: true,
+          }),
+        ).unwrap();
+
+        const getApprovedDongil = (idToApprove: number) => {
+          let found;
+          proposedDongils.map((proposedDongil) => {
+            found = proposedDongil.challengeList.find(
+              (challenge) => challenge.id === idToApprove,
+            );
+          });
+          return found;
+        };
+        const approvedDongil = getApprovedDongil(id)!;
+        dispatch(appendThisWeekSDongil({ selectedKid, approvedDongil }));
+        openApproveCompletedSheet();
+      } catch (error) {
+        console.log(error);
+      } finally {
+        setApproveProposedDongilStatus('idle');
+      }
+    }
+  }
+
+  // 2. 자녀의 돈길을 수락할까요?
+  const openApproveCheckSheet = () => {
+    setOpenBottomSheet({
+      sheetContent: 'ApproveCheck',
+      sheetProps: { open: true },
+      contentProps: {
+        onApproveButtonClick: handleApproveButtonClick,
+        onDismiss: setCloseBottomSheet,
+      },
+    });
+  };
+
+  // 4. 자녀의 돈길이 수락되었어요
+  const openApproveCompletedSheet = () => {
+    const openSheet = () =>
+      setOpenBottomSheet({
+        sheetContent: 'SheetCompleted',
+        sheetProps: { open: true },
+        contentProps: {
+          type: 'approve',
+          onDismiss: setCloseBottomSheet,
+        },
+      });
+    openSheetBySequence(openSheet);
+  };
+
+  // 1. 제안받은 돈길 모달
+  const { openModal } = useModals();
   const navigate = useNavigate();
   function openProposedReceiptModal() {
     openModal(modals.receiptModal, {
@@ -38,19 +119,17 @@ function ProposedDongilItem({
         navigate(`/reject/${id}`);
       },
       onExtraSubmit: () => {
-        setIdToApprove!(id);
-        onApproveCheckOpen!();
+        openApproveCheckSheet!();
       },
-      isKid: false,
-      createdAt: createdAt,
-      interestRate: interestRate,
-      isMom: isMom,
-      itemName: itemName,
-      title: title,
-      totalPrice: totalPrice,
-      weekPrice: weekPrice,
-      weeks: weeks,
-      filename: 'dummy',
+      createdAt,
+      interestRate,
+      itemName,
+      title,
+      totalPrice,
+      weekPrice,
+      weeks,
+      fileName,
+      isMom,
       shouldCloseOnOverlayClick: true,
     });
   }
